vim-patch:8.1.1957: more code can be moved to evalvars.c
Problem:    More code can be moved to evalvars.c.
Solution:   Move code to where it fits better. (Yegappan Lakshmanan,
            closes vim/vim#4883)

https://github.com/vim/vim/commit/da6c03342117fb7f4a8110bd9e8627b612a05a64
